Simplify Neotree usage with LazyVim and monorepos
The new mappings are aligned to match LazyVim's default configuration
for NeoTree, as it's more natural to navigate.

The NeoTree shortcut is removed in favor of using NeoTree's default
`<leader>e` and `<leader>E` shortcuts instead.

Vim's current working directory is overriden as both LazyVim and LSP
Config cause the CWD to be interpreted differently when working in a
monorepo, blocking navigation in NeoTree and Telescope across the entire
repo.